Added MenuForm And ScoreForm
Added the menuform and the scoreform with all their resoureces. The functionality of the scoreform is not yet implemented(I will have to connect it to a database) and also I need to fix the bug when the user closes the game form or the score form from the X button and the application does not terminate.